+++ src/main/java/svenhjol/charm/decoration/module/RandomAnimalTextures.java
@@ -8,6 +8,7 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.fml.client.registry.RenderingRegistry;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import svenhjol.charm.Charm;
 import svenhjol.charm.base.CharmCategories;
 import svenhjol.charm.decoration.render.CustomWolfRenderer;
@@ -15,6 +16,7 @@ import svenhjol.meson.MesonModule;
 import svenhjol.meson.iface.IMesonEnum;
 import svenhjol.meson.iface.Module;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
@@ -22,14 +24,25 @@ import java.util.UUID;
 @Module(mod = Charm.MOD_ID, category = CharmCategories.DECORATION,
     description = "Wolves will spawn with texture variants.")
 public class RandomAnimalTextures extends MesonModule {
-    public static List<String> wolves = Arrays.asList(
-        "minecraft:wolf",
-        "charm:brownwolf",
-        "charm:greywolf",
-        "charm:blackwolf",
-        "charm:amotwolf",
-        "charm:jupiter1390"
-    );
+    public static List<String> wolves = new ArrayList<>();
+
+    @Override
+    public void onCommonSetup(FMLCommonSetupEvent event) {
+        // standard textures
+        wolves.addAll(Arrays.asList(
+            "minecraft:wolf",
+            "charm:brownwolf",
+            "charm:greywolf",
+            "charm:blackwolf",
+            "charm:amotwolf",
+            "charm:jupiter1390"
+        ));
+
+        // add NeverLoseGuy textures
+        for (int i = 1; i <= 25; i++) {
+            wolves.add("charm:wolf" + i);
+        }
+    }
 
     public enum MobType implements IMesonEnum {
         WOLF